add customHandleSummary to all k6 tests
