+++ k6/crds/update_destructive_crds.js
@@ -20,6 +20,8 @@ export const epDataRecv = new Trend('endpoint_data_recv');
 export const headerDataRecv = new Trend('header_data_recv');
 export const timePolled = new Trend('time_polled', true);
 
+export const handleSummary = k6Util.customHandleSummary;
+
 export const options = {
   scenarios: {
     destructiveUpdate: {
